refactor(sales-app): use BigDecimals whenever the user is dealing with price

+++ sales-app/app/src/main/java/com/peyess/salesapp/data/adapter/edit_service_order/frames/EditFramesDataEntityAdapter.kt
@@ -2,6 +2,7 @@ package com.peyess.salesapp.data.adapter.edit_service_order.frames
 
 import com.peyess.salesapp.data.model.edit_service_order.frames.EditFramesDataEntity
 import com.peyess.salesapp.data.model.local_sale.frames.LocalFramesDocument
+import java.math.BigDecimal
 
 fun EditFramesDataEntity.toLocalFramesDocument(): LocalFramesDocument {
     return LocalFramesDocument(
@@ -9,7 +10,7 @@ fun EditFramesDataEntity.toLocalFramesDocument(): LocalFramesDocument {
         areFramesNew = areFramesNew,
         design = description,
         reference = reference,
-        value = value,
+        value = value.toBigDecimal(),
         tagCode = tagCode,
         type = type,
         framesInfo = framesInfo,
