+++ sales-app/app/src/main/java/com/peyess/salesapp/data/adapter/local_sale/payment/SalePaymentEntityAdapter.kt
@@ -2,6 +2,7 @@ package com.peyess.salesapp.data.adapter.local_sale.payment
 
 import com.peyess.salesapp.data.model.local_sale.payment.LocalPaymentDocument
 import com.peyess.salesapp.data.model.local_sale.payment.LocalPaymentEntity
+import java.math.BigDecimal
 
 fun LocalPaymentEntity.toLocalPaymentDocument(): LocalPaymentDocument {
     return LocalPaymentDocument(
@@ -15,7 +16,7 @@ fun LocalPaymentEntity.toLocalPaymentDocument(): LocalPaymentDocument {
         methodId = methodId,
         methodName = methodName,
         methodType = methodType,
-        value = value,
+        value = value.toBigDecimal(),
         installments = installments,
         document = document,
