+++ sales-app/app/src/main/java/com/peyess/salesapp/data/adapter/product_sold_desc/FSLensSoldDescriptionAdapter.kt
@@ -12,7 +12,7 @@ fun FSLensSoldDescription.toLensSoldDescriptionDocument(): LensSoldDescriptionDo
         id = id,
         units = units,
         nameDisplay = nameDisplay,
-        price = price,
+        price = price.toBigDecimal(),
         discount = discount.toDiscountDescriptionDocument(),
 
         isDiscounted = isDiscounted,
