+++ sales-app/app/src/main/java/com/peyess/salesapp/screen/sale/payment/adapter/PaymentFeeDocumentAdapter.kt
@@ -7,6 +7,6 @@ fun PaymentFeeDocument.toPaymentFee(): PaymentFee {
     return PaymentFee(
         saleId = saleId,
         method = method,
-        value = value.toDouble(),
+        value = value,
     )
 }
